fichiers_exercices/5-moderation/requirements.txt
load-dotenv==0.1.0
openai==0.27.4
python-dotenv==1.0.0
requests==2.28.2

